wip: some scaffolding
