+++ dev_requirements.txt
@@ -8,13 +8,14 @@ appdirs==1.4.3            # via black
 attrs==19.3.0             # via black, flake8-bugbear, pytest, syrupy
 black==19.10b0            # via syrupy
 bleach==3.1.4             # via readme-renderer
-certifi==2020.4.5         # via requests
+certifi==2020.4.5.1       # via requests
 chardet==3.0.4            # via requests
 click==7.1.1              # via black, pip-tools
 codecov==2.0.22           # via syrupy
 colored==1.4.2            # via syrupy
 configparser==5.0.0       # via py-githooks
 coverage[toml]==5.0.4     # via codecov, syrupy
+deprecated==1.2.8         # via pygithub
 docutils==0.16            # via readme-renderer
 entrypoints==0.3          # via flake8
 flake8-bugbear==20.1.4    # via syrupy
@@ -39,14 +40,16 @@ py-githooks==1.1.0        # via syrupy
 py==1.8.1                 # via pytest
 pycodestyle==2.5.0        # via flake8
 pyflakes==2.1.1           # via flake8
+pygithub==1.47            # via syrupy
 pygments==2.6.1           # via readme-renderer
-pyparsing==2.4.6          # via packaging
+pyjwt==1.7.1              # via pygithub
+pyparsing==2.4.7          # via packaging
 pyperf==2.0.0             # via syrupy
 pytest==5.4.1             # via syrupy
 readme-renderer==25.0     # via twine
 regex==2020.4.4           # via black
 requests-toolbelt==0.9.1  # via twine
-requests==2.23.0          # via codecov, requests-toolbelt, twine
+requests==2.23.0          # via codecov, pygithub, requests-toolbelt, twine
 semver==2.9.1             # via syrupy
 six==1.14.0               # via bleach, packaging, pip-tools, readme-renderer
 toml==0.10.0              # via black, coverage
@@ -58,6 +61,7 @@ urllib3==1.25.8           # via requests
 wcwidth==0.1.9            # via pytest
 webencodings==0.5.1       # via bleach
 wheel==0.34.2             # via syrupy
+wrapt==1.12.1             # via deprecated
 
 # The following packages are considered to be unsafe in a requirements file:
 setuptools==46.1.3        # via twine
